Fix (MSVC) compiler warnings

+++ src/framework/mpe/tests/utils/articulationutils.h
@@ -40,7 +40,7 @@ inline PitchPattern createSimplePitchPattern(const pitch_level_t incrementDiff =
 {
     PitchPattern result;
 
-    for (size_t i = 0; i < EXPECTED_SIZE; ++i) {
+    for (int i = 0; i < EXPECTED_SIZE; ++i) {
         result.pitchOffsetMap.insert_or_assign(i * TEN_PERCENTS, i * incrementDiff);
     }
 
@@ -53,7 +53,7 @@ inline ExpressionPattern createSimpleExpressionPattern(const dynamic_level_t amp
 
     float amplitudeSqrt = std::sqrt(amplitudeLevel);
 
-    for (size_t i = 0; i < EXPECTED_SIZE; ++i) {
+    for (int i = 0; i < EXPECTED_SIZE; ++i) {
         duration_percentage_t currentPos = i * TEN_PERCENTS;
         dynamic_level_t value = amplitudeLevel - std::pow(
             (2 * (amplitudeSqrt / static_cast<float>(HUNDRED_PERCENTS)) * currentPos) - amplitudeSqrt, 2);
